Revert "Move qt_for_kernel.py from external to lib"
This reverts commit d6765653cb16adf0fda050d3bb6d915fe67557c9.
