+++ IPython/external/qt_for_kernel.py
